improved createNewKafkaInstance
